fix: Moves some polls logic to middleware and use CONFERENCE_JOIN_IN_PROGRESS.
