+++ react/features/app/middlewares.any.js
@@ -37,7 +37,6 @@ import '../lobby/middleware';
 import '../notifications/middleware';
 import '../overlay/middleware';
 import '../polls/middleware';
-import '../polls/subscriber';
 import '../reactions/middleware';
 import '../recent-list/middleware';
 import '../recording/middleware';
